requirements.txt
py-pde
matplotlib==3.8.1
numpy
opencv-python

